refactor: split tag management functions
